* master: (62 commits)
  authpdo - use type safe comparison on passwords. fixes #1765
  removed unneeded files from random_compat
  Use cryptographically secure pseudo random number generator (CSPRNG)
  translation update
  translation update
  translation update
  translation update
  translation update
  or maybe I only now figured out the right travis config :-/
  disable 7.1 testing completely
  allow 7.1 failures until it's released #1682
  DE translation for key 'protected'
  Use 'Benachrichtigung' instead of 'Notifikation'
  Fix spelling of some terms (lower to upper case)
  DE translation for key: addUser_error_missing_pass
  DE translation for keys: update_name, update_mail
  translation update
  translation update
  Implemented interwiki substitution for external images (issue #1614).
  translation update
  ...

+++ inc/template.php
@@ -1224,8 +1224,8 @@ function tpl_img($maxwidth = 0, $maxheight = 0, $link = true, $params = null) {
     /** @var Input $INPUT */
     global $INPUT;
     global $REV;
-    $w = tpl_img_getTag('File.Width');
-    $h = tpl_img_getTag('File.Height');
+    $w = (int) tpl_img_getTag('File.Width');
+    $h = (int) tpl_img_getTag('File.Height');
 
     //resize to given max values
     $ratio = 1;
